Made a shell script!
Matches the batch file in functionality, but the structure uses a lot less goto's in favor of better if statements. This file has the added feature of a 'setup' function that will use gcc and g++ to compile the C and C++. I will look into adding that to the batch file, but there really isn't a standard compiler on Windows.